address PR comments and fix consistency issues
