+++ examples/advanced/experiment-tracking/mlflow/jobs/hello-pt-mlflow-client/code/src/client.py
@@ -19,43 +19,21 @@ import torch.nn as nn
 import torch.optim as optim
 import torchvision
 import torchvision.transforms as transforms
-from net import Net
+from model import SimpleNetwork
 
 # (1) import nvflare client API
 import nvflare.client as flare
 from nvflare.app_common.app_constant import ModelName
-from nvflare.client.tracking import WandBWriter
+
+# (2) import nvflare MLflowWriter
+from nvflare.client.tracking import MLflowWriter
 
 # (optional) set a fix place so we don't need to download everytime
 CIFAR10_ROOT = "/tmp/nvflare/data/cifar10"
+
 # (optional) We change to use GPU to speed things up.
 # if you want to use CPU, change DEVICE="cpu"
-DEVICE = "cuda:0" if torch.cuda.is_available() else "cpu"
-
-
-# wraps evaluation logic into a method to re-use for
-#       evaluation on both trained and received model
-def evaluate(input_weights, data_loader):
-    net = Net()
-    net.load_state_dict(input_weights)
-    # (optional) use GPU to speed things up
-    net.to(DEVICE)
-
-    correct = 0
-    total = 0
-    # since we're not training, we don't need to calculate the gradients for our outputs
-    with torch.no_grad():
-        for data in data_loader:
-            # (optional) use GPU to speed things up
-            images, labels = data[0].to(DEVICE), data[1].to(DEVICE)
-            # calculate outputs by running images through the network
-            outputs = net(images)
-            # the class with the highest energy is what we choose as prediction
-            _, predicted = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-
-    return 100 * correct // total
+DEVICE = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
 
 def define_parser():
@@ -83,18 +61,44 @@ def main():
     trainloader = torch.utils.data.DataLoader(trainset, batch_size=batch_size, shuffle=True, num_workers=num_workers)
     testset = torchvision.datasets.CIFAR10(root=dataset_path, train=False, download=True, transform=transform)
     testloader = torch.utils.data.DataLoader(testset, batch_size=batch_size, shuffle=False, num_workers=num_workers)
+    n_loaders = len(trainloader)
 
-    net = Net()
+    net = SimpleNetwork()
     best_accuracy = 0.0
 
-    # (2) initialize NVFlare client API
-    flare.init()
-    wandb_writer = WandBWriter()
+    # wraps evaluation logic into a method to re-use for
+    #       evaluation on both trained and received model
+    def evaluate(input_weights):
+        net = SimpleNetwork()
 
-    # (3) run continuously when launch_once=true
+        net.load_state_dict(input_weights)
+        # (optional) use GPU to speed things up
+        net.to(DEVICE)
+
+        correct = 0
+        total = 0
+        # since we're not training, we don't need to calculate the gradients for our outputs
+        with torch.no_grad():
+            for data in testloader:
+                # (optional) use GPU to speed things up
+                images, labels = data[0].to(DEVICE), data[1].to(DEVICE)
+                # calculate outputs by running images through the network
+                outputs = net(images)
+                # the class with the highest energy is what we choose as prediction
+                _, predicted = torch.max(outputs.data, 1)
+                total += labels.size(0)
+                correct += (predicted == labels).sum().item()
+
+        return 100 * correct // total
+
+    # (3) initialize NVFlare client API
+    flare.init()
+    mlflow_writer = MLflowWriter()
+
+    # (4) run continuously when launch_once=true
     while flare.is_running():
 
-        # (4) receive FLModel from NVFlare
+        # (5) receive FLModel from NVFlare
         input_model = flare.receive()
         client_id = flare.get_site_name()
 
@@ -105,11 +109,11 @@ def main():
         # for "evaluate" task (flare.is_evaluate()) we use the received model to do evaluation
         # and send back the evaluation metrics
         # for "submit_model" task (flare.is_submit_model()) we just need to send back the local model
-        # (5) performing train task on received model
+        # (6) performing train task on received model
         if flare.is_train():
             print(f"({client_id}) current_round={input_model.current_round}, total_rounds={input_model.total_rounds}")
 
-            # (5.1) loads model from NVFlare
+            # (6.1) loads model from NVFlare
             net.load_state_dict(input_model.params)
 
             criterion = nn.CrossEntropyLoss()
@@ -117,7 +121,7 @@ def main():
 
             # (optional) use GPU to speed things up
             net.to(DEVICE)
-            # (optional) calculate total steps for each round
+            # (optional) calculate total steps
             steps = local_epochs * len(trainloader)
             for epoch in range(local_epochs):  # loop over the dataset multiple times
 
@@ -138,48 +142,54 @@ def main():
 
                     # print statistics
                     running_loss += loss.item()
-                    if i % 3000 == 2999:  # print every 3000 mini-batches
-                        print(f"({client_id}) [{epoch + 1}, {i + 1:5d}] loss: {running_loss / 3000:.3f}")
-                        global_step = input_model.current_round * steps + epoch * len(trainloader) + i
-                        wandb_writer.log(metrics={"train_loss": running_loss / 3000}, step=global_step)
+                    if i % 2000 == 1999:  # print every 2000 mini-batches
+                        print(f"({client_id}) [{epoch + 1}, {i + 1:5d}] loss: {running_loss / 2000:.3f}")
                         running_loss = 0.0
+                        break
 
             print(f"({client_id}) Finished Training")
 
-            # (5.2) evaluation on local trained model to save best model
-            local_accuracy = evaluate(net.state_dict(), testloader)
+            # (6.2) evaluation on local trained model to save best model
+            local_accuracy = evaluate(net.state_dict())
+            global_step = input_model.current_round * n_loaders + i
+
+            # log metrics to MLflow
+            mlflow_writer.log_metric(key="local_accuracy", value=local_accuracy, step=global_step)
             print(f"({client_id}) Evaluating local trained model. Accuracy on the 10000 test images: {local_accuracy}")
             if local_accuracy > best_accuracy:
                 best_accuracy = local_accuracy
                 torch.save(net.state_dict(), model_path)
 
-            # (5.3) evaluate on received model for model selection
-            accuracy = evaluate(input_model.params, testloader)
+            # (6.3) evaluate on received model for model selection
+            accuracy = evaluate(input_model.params)
             print(
                 f"({client_id}) Evaluating received model for model selection. Accuracy on the 10000 test images: {accuracy}"
             )
 
-            # (5.4) construct trained FL model
+            # (6.4) construct trained FL model
             output_model = flare.FLModel(
                 params=net.cpu().state_dict(),
                 metrics={"accuracy": accuracy},
                 meta={"NUM_STEPS_CURRENT_ROUND": steps},
             )
 
-            # (5.5) send model back to NVFlare
+            # (6.5) send model back to NVFlare
             flare.send(output_model)
-        # (6) performing evaluate task on received model
+
+        # (7) performing evaluate task on received model
         elif flare.is_evaluate():
-            accuracy = evaluate(input_model.params, testloader)
+            accuracy = evaluate(input_model.params)
+            print(f"({client_id}) accuracy: {accuracy}")
             flare.send(flare.FLModel(metrics={"accuracy": accuracy}))
 
-        # (7) performing submit_model task to return the best local model
+        # (8) performing submit_model task to obtain best local model
         elif flare.is_submit_model():
             model_name = input_model.meta["submit_model_name"]
             if model_name == ModelName.BEST_MODEL:
                 try:
                     weights = torch.load(model_path)
-                    net = Net()
+                    net = SimpleNetwork()
+
                     net.load_state_dict(weights)
                     flare.send(flare.FLModel(params=net.cpu().state_dict()))
                 except Exception as e:
