+++ examples/advanced/experiment-tracking/mlflow/jobs/hello-pt-mlflow/code/job.py
@@ -14,7 +14,7 @@
 
 import argparse
 
-from src.network import SimpleNetwork
+from src.model import SimpleNetwork
 
 from nvflare.app_opt.pt.recipes import FedAvgRecipe
 from nvflare.recipe.utils import add_experiment_tracking
@@ -45,7 +45,7 @@ if __name__ == "__main__":
         min_clients=args.n_clients,
         num_rounds=5,
         initial_model=SimpleNetwork(),
-        train_script="src/train_script.py",
+        train_script="src/client.py",
     )
 
     # Add MLflow tracking
