+++ examples/advanced/experiment-tracking/tensorboard/jobs/tensorboard-streaming/code/job.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from src.network import SimpleNetwork
+from src.model import SimpleNetwork
 
 from nvflare.app_opt.pt.recipes import FedAvgRecipe
 from nvflare.recipe.utils import add_experiment_tracking
@@ -24,7 +24,7 @@ if __name__ == "__main__":
         min_clients=2,
         num_rounds=5,
         initial_model=SimpleNetwork(),
-        train_script="src/train_script.py",
+        train_script="src/client.py",
     )
 
     # Add TensorBoard tracking
