+++ examples/advanced/experiment-tracking/tensorboard/jobs/tensorboard-streaming/code/src/client.py
@@ -19,7 +19,7 @@ import torch.nn as nn
 import torch.optim as optim
 import torchvision
 import torchvision.transforms as transforms
-from network import SimpleNetwork
+from model import SimpleNetwork
 
 # (1) import nvflare client API
 import nvflare.client as flare
